fixed bugs with pick up
