+++ DoodleMyNoodle/Assets/Code/Game/Gameplay/Presentation/HUD/Items/InteractableInventoryDisplay.cs.meta
